Use shouldShowJoinLeaveMessages everywhere
