+++ webapp/channels/src/components/post_view/combined_system_message/index.ts
@@ -8,8 +8,7 @@ import type {Dispatch} from 'redux';
 import type {GlobalState} from '@mattermost/types/store';
 
 import {getMissingProfilesByIds, getMissingProfilesByUsernames} from 'mattermost-redux/actions/users';
-import {Preferences} from 'mattermost-redux/constants';
-import {getBool} from 'mattermost-redux/selectors/entities/preferences';
+import {shouldShowJoinLeaveMessages} from 'mattermost-redux/selectors/entities/preferences';
 import {getCurrentUser, makeGetProfilesByIdsAndUsernames} from 'mattermost-redux/selectors/entities/users';
 
 import CombinedSystemMessage from './combined_system_message';
@@ -29,7 +28,7 @@ function makeMapStateToProps() {
         return {
             currentUserId: currentUser.id,
             currentUsername: currentUser.username,
-            showJoinLeave: getBool(state, Preferences.CATEGORY_ADVANCED_SETTINGS, Preferences.ADVANCED_FILTER_JOIN_LEAVE, true),
+            showJoinLeave: shouldShowJoinLeaveMessages(state),
             userProfiles: getProfilesByIdsAndUsernames(state, {allUserIds, allUsernames}),
         };
     };
