+++ webapp/channels/src/packages/mattermost-redux/src/selectors/entities/posts.ts
@@ -24,11 +24,11 @@ import {createSelector} from 'mattermost-redux/selectors/create_selector';
 import {getChannel} from 'mattermost-redux/selectors/entities/channels';
 import {getCurrentChannelId, getCurrentUser} from 'mattermost-redux/selectors/entities/common';
 import {getConfig} from 'mattermost-redux/selectors/entities/general';
-import {getBool, getMyPreferences} from 'mattermost-redux/selectors/entities/preferences';
+import {getBool, shouldShowJoinLeaveMessages} from 'mattermost-redux/selectors/entities/preferences';
 import {getCurrentTeamId} from 'mattermost-redux/selectors/entities/teams';
 import {getUsers, getCurrentUserId, getUserStatuses} from 'mattermost-redux/selectors/entities/users';
 import {createIdsSelector} from 'mattermost-redux/utils/helpers';
-import {isCombinedUserActivityPost, shouldShowJoinLeaveMessages} from 'mattermost-redux/utils/post_list';
+import {isCombinedUserActivityPost} from 'mattermost-redux/utils/post_list';
 import {
     isPostEphemeral,
     isSystemMessage,
@@ -37,7 +37,6 @@ import {
     isPostPendingOrFailed,
     isPostCommentMention,
 } from 'mattermost-redux/utils/post_utils';
-import {getPreferenceKey} from 'mattermost-redux/utils/preference_utils';
 import {isGuest} from 'mattermost-redux/utils/user_utils';
 
 export function getAllPosts(state: GlobalState) {
@@ -345,17 +344,14 @@ export function makeGetPostsInChannel(): (state: GlobalState, channelId: Channel
         getPostsInThread,
         (state: GlobalState, channelId: Channel['id']) => getPostIdsInChannel(state, channelId),
         getCurrentUser,
-        getMyPreferences,
+        shouldShowJoinLeaveMessages,
         (state: GlobalState, channelId: Channel['id'], numPosts: number) => numPosts || Posts.POST_CHUNK_SIZE,
-        (allPosts, postsInThread, allPostIds, currentUser, myPreferences, numPosts) => {
+        (allPosts, postsInThread, allPostIds, currentUser, showJoinLeave, numPosts) => {
             if (!allPostIds) {
                 return null;
             }
 
             const posts: PostWithFormatData[] = [];
-
-            const joinLeavePref = myPreferences[getPreferenceKey(Preferences.CATEGORY_ADVANCED_SETTINGS, Preferences.ADVANCED_FILTER_JOIN_LEAVE)];
-            const showJoinLeave = joinLeavePref ? joinLeavePref.value !== 'false' : true;
 
             const postIds = numPosts === -1 ? allPostIds : allPostIds.slice(0, numPosts);
 
@@ -389,15 +385,13 @@ export function makeGetPostsAroundPost(): (state: GlobalState, postId: Post['id'
         getPostsInThread,
         (state: GlobalState, focusedPostId) => focusedPostId,
         getCurrentUser,
-        getMyPreferences,
-        (postIds, allPosts, postsInThread, focusedPostId, currentUser, myPreferences) => {
+        shouldShowJoinLeaveMessages,
+        (postIds, allPosts, postsInThread, focusedPostId, currentUser, showJoinLeave) => {
             if (!postIds || !currentUser) {
                 return null;
             }
 
             const posts: PostWithFormatData[] = [];
-            const joinLeavePref = myPreferences[getPreferenceKey(Preferences.CATEGORY_ADVANCED_SETTINGS, Preferences.ADVANCED_FILTER_JOIN_LEAVE)];
-            const showJoinLeave = joinLeavePref ? joinLeavePref.value !== 'false' : true;
 
             for (let i = 0; i < postIds.length; i++) {
                 const post = allPosts[postIds[i]];
@@ -566,13 +560,11 @@ export const getMostRecentPostIdInChannel: (state: GlobalState, channelId: Chann
     'getMostRecentPostIdInChannel',
     getAllPosts,
     (state: GlobalState, channelId: string) => getPostIdsInChannel(state, channelId),
-    getMyPreferences,
-    (posts, postIdsInChannel, preferences) => {
+    shouldShowJoinLeaveMessages,
+    (posts, postIdsInChannel, allowSystemMessages) => {
         if (!postIdsInChannel) {
             return '';
         }
-        const key = getPreferenceKey(Preferences.CATEGORY_ADVANCED_SETTINGS, Preferences.ADVANCED_FILTER_JOIN_LEAVE);
-        const allowSystemMessages = preferences[key] ? preferences[key].value === 'true' : true;
 
         if (!allowSystemMessages) {
             // return the most recent non-system message in the channel
