+++ webapp/channels/src/packages/mattermost-redux/src/selectors/entities/preferences.test.ts
@@ -599,6 +599,98 @@ describe('Selectors.Preferences', () => {
     });
 });
 
+describe('shouldShowJoinLeaveMessages', () => {
+    it('should default to true', () => {
+        const state = {
+            entities: {
+                general: {
+                    config: {
+                        EnableJoinLeaveMessageByDefault: 'true',
+                    },
+                },
+                preferences: {
+                    myPreferences: {},
+                },
+            },
+        } as unknown as GlobalState;
+
+        // Defaults to show post
+        const show = Selectors.shouldShowJoinLeaveMessages(state);
+        expect(show).toEqual(true);
+    });
+
+    it('set config to false, return false', () => {
+        const state = {
+            entities: {
+                general: {
+                    config: {
+                        EnableJoinLeaveMessageByDefault: 'false',
+                    },
+                },
+                preferences: {
+                    myPreferences: {},
+                },
+            },
+        } as unknown as GlobalState;
+
+        // Defaults to show post
+        const show = Selectors.shouldShowJoinLeaveMessages(state);
+        expect(show).toEqual(false);
+    });
+
+    it('if user preference, set default wont be used', () => {
+        const state = {
+            entities: {
+                general: {
+                    config: {
+                        EnableJoinLeaveMessageByDefault: 'false',
+                    },
+                },
+                preferences: {
+                    myPreferences: {
+                        [getPreferenceKey(Preferences.CATEGORY_ADVANCED_SETTINGS, Preferences.ADVANCED_FILTER_JOIN_LEAVE)]: {
+                            category: Preferences.CATEGORY_ADVANCED_SETTINGS,
+                            name: Preferences.ADVANCED_FILTER_JOIN_LEAVE,
+                            value: 'true',
+                        },
+
+                    },
+                },
+            },
+        } as unknown as GlobalState;
+
+        // Defaults to show post
+        const show = Selectors.shouldShowJoinLeaveMessages(state);
+        expect(show).toEqual(true);
+    });
+
+    it('if user preference, set default wont be used', () => {
+        const state = {
+            entities: {
+                general: {
+                    config: {
+                        EnableJoinLeaveMessageByDefault: 'true',
+                    },
+                },
+                preferences: {
+                    myPreferences: {
+                        [getPreferenceKey(Preferences.CATEGORY_ADVANCED_SETTINGS, Preferences.ADVANCED_FILTER_JOIN_LEAVE)]: {
+                            category: Preferences.CATEGORY_ADVANCED_SETTINGS,
+                            name: Preferences.ADVANCED_FILTER_JOIN_LEAVE,
+                            value: 'false',
+                        },
+
+                    },
+                },
+            },
+        } as unknown as GlobalState;
+
+        // Defaults to show post
+        const show = Selectors.shouldShowJoinLeaveMessages(state);
+        expect(show).toEqual(false);
+    });
+});
+
 describe('shouldShowUnreadsCategory', () => {
     test('should return value from the preference if set', () => {
         const state = {
