+++ webapp/channels/src/packages/mattermost-redux/src/utils/post_list.ts
@@ -6,12 +6,11 @@ import moment from 'moment-timezone';
 import type {ActivityEntry, Post} from '@mattermost/types/posts';
 import type {GlobalState} from '@mattermost/types/store';
 
-import {Posts, Preferences} from 'mattermost-redux/constants';
+import {Posts} from 'mattermost-redux/constants';
 import {createSelector} from 'mattermost-redux/selectors/create_selector';
-import {getConfig} from 'mattermost-redux/selectors/entities/general';
 import {makeGetPostsForIds} from 'mattermost-redux/selectors/entities/posts';
 import type {UserActivityPost} from 'mattermost-redux/selectors/entities/posts';
-import {getBool} from 'mattermost-redux/selectors/entities/preferences';
+import {shouldShowJoinLeaveMessages} from 'mattermost-redux/selectors/entities/preferences';
 import {getCurrentUser} from 'mattermost-redux/selectors/entities/users';
 import {createIdsSelector, memoizeResult} from 'mattermost-redux/utils/helpers';
 import {isUserActivityPost, shouldFilterJoinLeavePost, isFromWebhook} from 'mattermost-redux/utils/post_utils';
@@ -22,14 +21,6 @@ export const CREATE_COMMENT = 'create-comment';
 export const DATE_LINE = 'date-';
 export const START_OF_NEW_MESSAGES = 'start-of-new-messages-';
 export const MAX_COMBINED_SYSTEM_POSTS = 100;
-
-export function shouldShowJoinLeaveMessages(state: GlobalState) {
-    const config = getConfig(state);
-    const enableJoinLeaveMessage = config.EnableJoinLeaveMessageByDefault === 'true';
-
-    // This setting is true or not set if join/leave messages are to be displayed
-    return getBool(state, Preferences.CATEGORY_ADVANCED_SETTINGS, Preferences.ADVANCED_FILTER_JOIN_LEAVE, enableJoinLeaveMessage);
-}
 
 interface PostFilterOptions {
     postIds: string[];
